refactor: 디렉토리 및 import문 수정 - user 관련 인수테스트 디렉토리 이동 - 빠진 import문 추가
