+++ backend/src/test/java/kr/kro/colla/user/user/AcceptanceTest.java
@@ -1,0 +1,88 @@
+package kr.kro.colla.user.user;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import kr.kro.colla.auth.service.JwtProvider;
+import kr.kro.colla.common.fixture.User;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.notNullValue;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+    @LocalServerPort
+    int port;
+
+    @Autowired
+    private JwtProvider jwtProvider;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+        user = new User(jwtProvider);
+    }
+
+    private Long managerId = 3L;
+    private String name = "프로젝트 이름", desc = "프로젝트 설명";
+    private User user;
+
+    @Test
+    void 사용자_프로젝트_생성_후_반환한다() {
+        // given
+        String accessToken = user.로그인(managerId);
+        Map<String, Object> requestBody = new HashMap<>();
+        requestBody.put("name", name);
+        requestBody.put("description",desc);
+
+        given()
+                .contentType(ContentType.JSON)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .cookie("accessToken", accessToken)
+                .body(requestBody)
+        // when
+        .when()
+                .post("/api/users/{userId}/projects", managerId)
+        // then
+        .then()
+                .statusCode(HttpStatus.OK.value())
+                .body("id", notNullValue())
+                .body("managerId", equalTo(managerId.intValue()))
+                .body("name", equalTo(name))
+                .body("description", equalTo(desc));
+    }
+
+    @Test
+    void 사용자_프로젝트_생성_실패_시_에러를_반환한다() throws Exception {
+        // given
+        String accessToken = user.로그인(managerId);
+        Map<String, Object> requestBody = new HashMap<>();
+        requestBody.put("description",desc);
+
+        given()
+                .contentType(ContentType.JSON)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .cookie("accessToken", accessToken)
+                .body(requestBody)
+        // when
+        .when()
+                .post("/api/users/{userId}/projects",managerId)
+
+        // then
+        .then()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .body("status", equalTo(HttpStatus.BAD_REQUEST.value()))
+                .body("message", equalTo("name : 공백일 수 없습니다"));
+    }
+}
